+++ bin/network-monitor/src/status.rs
@@ -177,6 +177,16 @@ pub struct NoteTransportStatusDetails {
     pub serving_status: String,
 }
 
+/// Details of the validator service.
+#[derive(Debug, Clone, Serialize, Deserialize, Default)]
+pub struct ValidatorStatusDetails {
+    pub url: String,
+    pub version: String,
+    pub chain_tip: u32,
+    pub validated_transactions_count: u64,
+    pub signed_blocks_count: u64,
+}
+
 /// Details of a service.
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub enum ServiceDetails {
@@ -188,6 +198,7 @@ pub enum ServiceDetails {
     NtxTracking(CounterTrackingDetails),
     ExplorerStatus(ExplorerStatusDetails),
     NoteTransportStatus(NoteTransportStatusDetails),
+    ValidatorStatus(ValidatorStatusDetails),
     Error,
 }